Move MetaWarning classes into new cli module
Moving them out of __main__.py will make it easier to re-use them across
different command-line interfaces, and may make them easier to test.

While moving them, I also added the missing docstrings.
